+++ inlang/packages/paraglide/paraglide-js/examples/astro/src/components/Counter.svelte
@@ -1,0 +1,9 @@
+<script>
+    import * as m from "../paraglide/messages.js";
+
+    let count = 0;
+</script>
+
+<button on:click={() => count += 1}>
+    {m.count({  count })}
+</button>