+++ inlang/packages/paraglide/paraglide-js/src/compiler/output-structure/message-modules.ts
@@ -2,9 +2,9 @@ import type { ProjectSettings } from "@inlang/sdk";
 import type { CompiledBundleWithMessages } from "../compile-bundle.js";
 import { createRuntimeFile } from "../runtime/create-runtime.js";
 import { createRegistry } from "../registry.js";
-import { jsIdentifier } from "../../services/codegen/identifier.js";
 import { escapeForSingleQuoteString } from "../../services/codegen/escape.js";
-import type { CompilerOptions } from "../compile.js";
+import type { CompilerOptions } from "../compiler-options.js";
+import { toSafeModuleId } from "../safe-module-id.js";
 
 export function generateMessageModules(
 	compiledBundles: CompiledBundleWithMessages[],
@@ -24,36 +24,49 @@ export function generateMessageModules(
 		["registry.js"]: createRegistry(),
 	};
 
-	// messages index file
-	output["messages.js"] = [
+	// all messages index file
+	output["messages/_index.js"] = [
 		...compiledBundles.map(
-			({ bundle }) => `export * from './messages/${bundle.node.id}/index.js'`
+			({ bundle }) =>
+				`export * from './${toSafeModuleId(bundle.node.id)}/index.js'`
 		),
 	].join("\n");
 
-	// Creates a per message index file
-	for (const compiledBundle of compiledBundles) {
-		const filename = `messages/${compiledBundle.bundle.node.id}/index.js`;
-		const code = [
-			settings.locales
-				.map(
-					(locale) =>
-						`import * as ${jsIdentifier(locale)} from "./${locale}.js"`
-				)
-				.join("\n"),
-			`import { getLocale } from '../../runtime.js'`,
-			"",
-			compiledBundle.bundle.code,
-		].join("\n");
-		output[filename] = code;
-	}
+	output["messages.js"] = [
+		"export * from './messages/_index.js'",
+		"// enabling auto-import by exposing all messages as m",
+		"export * as m from './messages/_index.js'",
+	].join("\n");
 
-	for (const locale of settings.locales) {
-		for (const compiledBundle of compiledBundles) {
+	for (const compiledBundle of compiledBundles) {
+		const bundleFileId = toSafeModuleId(compiledBundle.bundle.node.id);
+		// bundle file
+		const indexFilename = `messages/${bundleFileId}/index.js`;
+		if (output[indexFilename]) {
+			// bundle file already exists, need to append to it
+			output[indexFilename] += `\n${compiledBundle.bundle.code}`;
+		} else {
+			// create fresh bundle file
+			const code = [
+				settings.locales
+					.map(
+						(locale) =>
+							`import * as ${toSafeModuleId(locale)} from "./${locale}.js"`
+					)
+					.join("\n"),
+				`import { getLocale } from '../../runtime.js'`,
+				"",
+				compiledBundle.bundle.code,
+			].join("\n");
+			output[indexFilename] = code;
+		}
+
+		// message files
+		for (const locale of settings.locales) {
 			let file = "";
 
 			const compiledMessage = compiledBundle.messages[locale];
-			const id = jsIdentifier(compiledBundle.bundle.node.id);
+			const id = toSafeModuleId(compiledBundle.bundle.node.id);
 			if (!compiledMessage) {
 				// add fallback
 				const fallbackLocale = fallbackMap[locale];
@@ -62,7 +75,7 @@ export function generateMessageModules(
 					file += `\nexport { ${id} } from "./${fallbackLocale}.js"`;
 				} else {
 					// fallback to just the bundle id
-					file += `\nexport const ${id} = () => '${escapeForSingleQuoteString(
+					file += `\n/** @type {(inputs?: Record<string, never>) => string} */\nexport const ${id} = () => '${escapeForSingleQuoteString(
 						compiledBundle.bundle.node.id
 					)}'`;
 				}
@@ -70,13 +83,18 @@ export function generateMessageModules(
 				file += `\n${compiledMessage.code}`;
 			}
 
-			// Add the registry import to the message file
-			// if registry is used
-			if (file.includes("registry.")) {
-				file = `import * as registry from '../../registry.js'\n` + file;
-			}
+			if (output[`messages/${bundleFileId}/${locale}.js`]) {
+				// message file already exists, need to append to it
+				output[`messages/${bundleFileId}/${locale}.js`] += file;
+			} else {
+				// Add the registry import to the message file
+				// if registry is used
+				if (file.includes("registry.")) {
+					file = `import * as registry from '../../registry.js'\n` + file;
+				}
 
-			output[`messages/${compiledBundle.bundle.node.id}/${locale}.js`] = file;
+				output[`messages/${bundleFileId}/${locale}.js`] = file;
+			}
 		}
 	}
 	return output;