+++ inlang/packages/paraglide/paraglide-js/src/compiler/runtime/get-locale.js
@@ -1,9 +1,11 @@
 import { assertIsLocale } from "./assert-is-locale.js";
+import { isLocale } from "./is-locale.js";
 import {
 	baseLocale,
 	strategy,
 	TREE_SHAKE_COOKIE_STRATEGY_USED,
 	TREE_SHAKE_GLOBAL_VARIABLE_STRATEGY_USED,
+	TREE_SHAKE_PREFERRED_LANGUAGE_STRATEGY_USED,
 	TREE_SHAKE_URL_STRATEGY_USED,
 } from "./variables.js";
 import { extractLocaleFromCookie } from "./extract-locale-from-cookie.js";
@@ -14,7 +16,7 @@ import { extractLocaleFromUrl } from "./extract-locale-from-url.js";
  * strategy and avoid type errors.
  *
  * The implementation is overwritten
- * by \`defineGetLocale()\` and \`defineSetLocale()\`.
+ * by \`overwriteGetLocale()\` and \`defineSetLocale()\`.
  *
  * @type {Locale|undefined}
  */
@@ -36,6 +38,15 @@ export let getLocale = () => {
 	/** @type {string | undefined} */
 	let locale;
 
+	// if running in a server-side rendering context
+	// retrieve the locale from the async local storage
+	if (serverMiddlewareAsyncStorage) {
+		const locale = serverMiddlewareAsyncStorage?.getStore()?.locale;
+		if (locale) {
+			return locale;
+		}
+	}
+
 	for (const strat of strategy) {
 		if (TREE_SHAKE_COOKIE_STRATEGY_USED && strat === "cookie") {
 			locale = extractLocaleFromCookie();
@@ -53,6 +64,12 @@ export let getLocale = () => {
 			_locale !== undefined
 		) {
 			locale = _locale;
+		} else if (
+			TREE_SHAKE_PREFERRED_LANGUAGE_STRATEGY_USED &&
+			strat === "preferredLanguage" &&
+			typeof window !== "undefined"
+		) {
+			locale = negotiatePreferredLanguageFromNavigator();
 		}
 		// check if match, else continue loop
 		if (locale !== undefined) {
@@ -60,24 +77,52 @@ export let getLocale = () => {
 		}
 	}
 
-	throw new Error("No locale found. There is an error in your strategy.");
+	throw new Error(
+		"No locale found. Read the docs https://inlang.com/m/gerre34r/library-inlang-paraglideJs/errors#no-locale-found"
+	);
 };
 
 /**
- * Define the \`getLocale()\` function.
+ * Negotiates a preferred language from navigator.languages.
  *
- * Use this function to define how the locale is resolved. For example,
+ * @returns {string|undefined} The negotiated preferred language.
+ */
+function negotiatePreferredLanguageFromNavigator() {
+	if (!navigator?.languages?.length) {
+		return undefined;
+	}
+
+	const languages = navigator.languages.map((lang) => ({
+		fullTag: lang.toLowerCase(),
+		baseTag: lang.split("-")[0]?.toLowerCase(),
+	}));
+
+	for (const lang of languages) {
+		if (isLocale(lang.fullTag)) {
+			return lang.fullTag;
+		} else if (isLocale(lang.baseTag)) {
+			return lang.baseTag;
+		}
+	}
+
+	return undefined;
+}
+
+/**
+ * Overwrite the \`getLocale()\` function.
+ *
+ * Use this function to overwrite how the locale is resolved. For example,
  * you can resolve the locale from the browser's preferred language,
  * a cookie, env variable, or a user's preference.
  *
  * @example
- *   defineGetLocale(() => {
+ *   overwriteGetLocale(() => {
  *     // resolve the locale from a cookie. fallback to the base locale.
  *     return Cookies.get('locale') ?? baseLocale
  *   }
  *
  * @type {(fn: () => Locale) => void}
  */
-export const defineGetLocale = (fn) => {
+export const overwriteGetLocale = (fn) => {
 	getLocale = fn;
 };