+++ inlang/packages/sherlock/src/utilities/editor/editor-app/CHANGELOG.md
@@ -1,5 +1,17 @@
 # editor-app
 
+## 0.0.5
+
+### Patch Changes
+
+- @inlang/editor-component@3.0.0
+
+## 0.0.4
+
+### Patch Changes
+
+- @inlang/editor-component@2.0.2
+
 ## 0.0.3
 
 ### Patch Changes