+++ inlang/packages/ui-components/editor-component/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @inlang/message-bundle-component
 
+## 3.0.0
+
+### Patch Changes
+
+- Updated dependencies [c0b857a]
+- Updated dependencies [91ba4eb]
+  - @inlang/sdk@2.3.0
+
+## 2.0.2
+
+### Patch Changes
+
+- Updated dependencies [c53b1a9]
+  - @inlang/sdk@2.2.2
+
 ## 2.0.1
 
 ### Patch Changes