+++ packages/lix-plugin-csv/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @lix-js/plugin-csv
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies [c0b857a]
+  - @lix-js/sdk@0.4.1
+
 ## 0.1.2
 
 ### Patch Changes